[folly/logging] Fix mixed MSVC/clang-cl build
Projects built with clang-cl that link against folly/logging built by
vcpkg fail at link time.  folly/logging has a number of functions
that are only defined when __INCLUDE_LEVEL__ is defined.  MSVC does not
support this macro, clang-cl does, resulting in a handful of functions
that are declared but not defined.

This mostly came from fc507b26548882a8492de57abf0a8923ecf7709a.

The negative consequence of this change is that XlogFileScopeInfo will
unnessarily larger (a pointer and an uint32) on compilers that don't
support __INCLUDE_LEVEL__ (e.g MSVC).